+++ earth2studio/models/dx/__init__.py
@@ -24,7 +24,7 @@ from earth2studio.models.dx.corrdiff import (
     CorrDiffTaiwan,
 )
 from earth2studio.models.dx.corrdiff_cmip6 import CorrDiffCMIP6
-from earth2studio.models.dx.cosmo_downscaling import CosmoDownscaling  # noqa
+from earth2studio.models.dx.cosmo_downscaling import CosmoDownscaling
 from earth2studio.models.dx.derived import (
     DerivedRH,
     DerivedRHDewpoint,